requirements.txt
# Core dependencies
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# API client dependencies
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
tenacity>=8.2.0

# S3 and cloud dependencies
boto3>=1.34.0

# Progress and logging
tqdm>=4.66.0
rich>=13.7.0

# Evaluation dependencies
opencv-python>=4.8.0
pandas>=2.1.0

# Web interface (optional)
gradio>=4.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
